Alert - Backend (load service), save Kpi - added  some check on relations

+++ knowage/src/it/eng/spagobi/kpi/KpiService.java
@@ -18,43 +18,6 @@
 package it.eng.spagobi.kpi;
 
 import static it.eng.spagobi.tools.scheduler.utils.SchedulerUtilitiesV2.getJobTriggerInfo;
-
-import java.io.IOException;
-import java.text.MessageFormat;
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
-import java.util.TreeMap;
-
-import javax.script.ScriptEngine;
-import javax.script.ScriptEngineManager;
-import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.Response;
-
-import org.apache.log4j.Logger;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.quartz.JobExecutionException;
-
-import com.fasterxml.jackson.core.JsonGenerationException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-
 import it.eng.spago.error.EMFErrorSeverity;
 import it.eng.spago.error.EMFInternalError;
 import it.eng.spago.error.EMFUserError;
@@ -101,9 +64,45 @@ import it.eng.spagobi.utilities.exceptions.SpagoBIException;
 import it.eng.spagobi.utilities.exceptions.SpagoBIServiceException;
 import it.eng.spagobi.utilities.rest.RestUtilities;
 
+import java.io.IOException;
+import java.text.MessageFormat;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.TreeMap;
+
+import javax.script.ScriptEngine;
+import javax.script.ScriptEngineManager;
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.DELETE;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.QueryParam;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.Response;
+
+import org.apache.log4j.Logger;
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+import org.quartz.JobExecutionException;
+
+import com.fasterxml.jackson.core.JsonGenerationException;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 /**
  * @authors Salvatore Lupo (Salvatore.Lupo@eng.it)
- *
+ * 
  */
 @Path("/1.0/kpi")
 @ManageAuthorization
@@ -271,7 +270,7 @@ public class KpiService {
 	 * Executes a given query over a given datasource (dataSourceId) limited by maxItem param. It uses existing backend to retrieve data and metadata, but the
 	 * resulting json is lightened in order to give back something like this: {"columns": [{"name": "column_1", "label": "order_id"},...], "rows": [{"column_1":
 	 * "1"},...]}
-	 *
+	 * 
 	 * @param req
 	 * @return
 	 * @throws EMFUserError
@@ -485,8 +484,13 @@ public class KpiService {
 	public Response deleteKpi(@PathParam("id") Integer id, @PathParam("version") Integer version, @QueryParam("toBeVersioned") Boolean toBeVersioned,
 			@Context HttpServletRequest req) throws EMFUserError {
 		IKpiDAO dao = getKpiDAO(req);
-		dao.removeKpi(id, version, Boolean.TRUE.equals(toBeVersioned));
-		return Response.ok().build();
+		JSError jsError = checkKpiRel(dao, id, version);
+		if (!jsError.hasErrors()) {
+			dao.removeKpi(id, version, Boolean.TRUE.equals(toBeVersioned));
+			return Response.ok().build();
+		} else {
+			return Response.ok(jsError.toString()).build();
+		}
 	}
 
 	@DELETE
@@ -694,6 +698,47 @@ public class KpiService {
 
 	}
 
+	private JSError checkKpiRel(IKpiDAO dao, Integer id, Integer version) {
+		JSError jsError = new JSError();
+		// Check if any relation exists with Scorecard
+		List<Scorecard> scorecards = dao.listScorecardByKpi(id, version);
+		if (scorecards != null && !scorecards.isEmpty()) {
+			String scNames = "";
+			for (int i = 0; i < scorecards.size(); i++) {
+				if (i != 0) {
+					scNames += ", ";
+				}
+				scNames += scorecards.get(i).getName();
+			}
+			jsError.addErrorKey("newKpi.kpi.kpiIsUsedByScorecards", scNames);
+		}
+		// Check if any relation exists with Target
+		List<Target> targets = dao.listTargetByKpi(id, version);
+		if (targets != null && !targets.isEmpty()) {
+			String names = "";
+			for (int i = 0; i < targets.size(); i++) {
+				if (i != 0) {
+					names += ", ";
+				}
+				names += targets.get(i).getName();
+			}
+			jsError.addErrorKey("newKpi.kpi.kpiIsUsedByTargets", names);
+		}
+		// Check if any relation exists with Scheduler
+		List<KpiScheduler> schedulers = dao.listSchedulerByKpi(id, version);
+		if (schedulers != null && !schedulers.isEmpty()) {
+			String names = "";
+			for (int i = 0; i < schedulers.size(); i++) {
+				if (i != 0) {
+					names += ", ";
+				}
+				names += schedulers.get(i).getName();
+			}
+			jsError.addErrorKey("newKpi.kpi.kpiIsUsedBySchedulers", names);
+		}
+		return jsError;
+	}
+
 	private JSError check(Target target, IKpiDAO dao) throws SpagoBIException {
 		JSError e = new JSError();
 		if (target.getName() == null) {
@@ -878,7 +923,7 @@ public class KpiService {
 
 	/**
 	 * Check if placeholders with default value are a subset of placeholders linked to measures used in kpi definition (ie kpi formula)
-	 *
+	 * 
 	 * @param servlet
 	 *            request
 	 * @param placeholder
@@ -1092,4 +1137,5 @@ public class KpiService {
 		}
 		return msg;
 	}
+
 }
